requirements.txt
discord.py
python-dotenv
requests
google-genai
pyyaml
PyMuPDF
